integrate bill to ethereum and bitcoin service

+++ bitcoin/src/main/java/com/github/bitcoin/entity/Account.java
@@ -152,7 +152,7 @@ public class Account implements Serializable, Cloneable {
         this.timeStamp = timeStamp;
     }
 
-    public void incoming(Transaction transaction, Long value) {
+    public void incoming(Transaction transaction, BigInteger value) {
         addTransaction(transaction);
         addAmount(value);
     }
